requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
python-dotenv==1.0.1
spark-ai-python
zhipuai==2.1.5.20250106
qianfan==0.4.12.3
unstructured==0.16.23
pymupdf==1.25.3
markdown==3.7
streamlit==1.43.0
jieba==0.42.1